+++ 0.0.1/src/auth_proxy.rs
@@ -1,50 +1,39 @@
 use std::str::FromStr;
 
+use actions::NearAction;
 use near_gas::NearGas;
 use near_sdk::collections::UnorderedSet;
 use near_sdk::ext_contract;
 use near_sdk::json_types::{Base58CryptoHash, U64};
 use near_sdk::{
-    bs58, env, near, AccountId, CurveType, Gas, NearToken, PanicOnDefault, Promise, PromiseError,
+    bs58, env, near, AccountId, Gas, NearToken, PanicOnDefault, Promise, PromiseError,
     PublicKey,
 };
 use near_sdk::base64;
 use hex::FromHex;
-use omni_transaction::near::types::Secp256K1PublicKey;
 use omni_transaction::transaction_builder::TransactionBuilder;
 use omni_transaction::transaction_builder::TxBuilder;
 use omni_transaction::near::near_transaction::NearTransaction;
 use omni_transaction::{
     near::types::{
         Action as OmniAction, BlockHash as OmniBlockHash,
-        FunctionCallAction as OmniFunctionCallAction, PublicKey as OmniPublicKey, U128 as OmniU128, Secp256K1Signature, Signature,
+        FunctionCallAction as OmniFunctionCallAction, U128 as OmniU128, Secp256K1Signature, Signature,
         U64 as OmniU64,
     },
     NEAR,
 };
-use sha2::{Digest, Sha256};
+
+mod models;
+mod actions;
+mod utils;
+
+pub use crate::models::*;
 
 #[ext_contract(ext_self)]
 pub trait ExtSelf {
     fn callback_method(&mut self, #[callback_result] call_result: Result<Vec<u8>, PromiseError>);
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct BigR {
-    pub affine_point: String,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ScalarValue {
-    pub scalar: String,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct SignatureResponse {
-    pub big_r: BigR,
-    pub s: ScalarValue,
-    pub recovery_id: u8,
-}
 
 const GAS_FOR_REQUEST_SIGNATURE: Gas = Gas::from_tgas(50);
 const BASE_GAS: Gas = Gas::from_tgas(5);  // Base gas for contract execution
@@ -59,25 +48,6 @@ pub struct ProxyContract {
     owner_id: AccountId,
     authorized_users: UnorderedSet<AccountId>,
     signer_contract: AccountId,
-}
-
-use borsh::{BorshDeserialize, BorshSerialize};
-use serde::{Deserialize, Serialize};
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize)]
-pub struct SignRequest {
-    pub payload: Vec<u8>,
-    pub path: String,
-    pub key_version: u32,
-}
-
-#[derive(Clone)]
-#[near(serializers = [json, borsh])]
-pub struct NearAction {
-    pub method_name: String,
-    pub contract_id: AccountId,
-    pub gas_attached: Gas,
-    pub deposit_attached: NearToken,
 }
 
 #[near]
@@ -101,6 +71,10 @@ impl ProxyContract {
         self.signer_contract.clone()
     }
 
+    pub fn get_owner_id(&self) -> AccountId {
+        self.owner_id.clone()
+    }
+
     #[payable]
     pub fn request_signature(
         &mut self,
@@ -112,6 +86,7 @@ impl ProxyContract {
         nonce: U64,
         block_hash: Base58CryptoHash,
         mpc_signer_pk: String,
+        account_pk_for_mpc: String,
     ) -> Promise {
 
         let attached_gas = env::prepaid_gas();
@@ -150,7 +125,7 @@ impl ProxyContract {
         };
 
         // verify the action is permitted
-        self.assert_action_allowed(&action);
+        NearAction::is_allowed(&action);
 
         let actions = vec![OmniAction::FunctionCall(Box::new(OmniFunctionCallAction {
             method_name: method_name.clone(),
@@ -164,7 +139,7 @@ impl ProxyContract {
         // construct the entire transaction to be signed
         let tx = TransactionBuilder::new::<NEAR>()
             .signer_id(env::current_account_id().to_string())
-            .signer_public_key(Self::convert_pk_to_omni(&PublicKey::from_str(&mpc_signer_pk).unwrap()))
+            .signer_public_key(utils::convert_pk_to_omni(&PublicKey::from_str(&mpc_signer_pk).unwrap()))
             .nonce(nonce.0) // Use the provided nonce
             .receiver_id(contract_id.to_string())
             .block_hash(OmniBlockHash(block_hash.into()))
@@ -193,19 +168,19 @@ impl ProxyContract {
         //https://github.com/near/near-api-js/blob/a33274d9c06fec7de756f4490dea0618b2fc75da/packages/providers/src/json-rpc-provider.ts#L112C32-L112C49
 
         // SHA-256 hash of the serialized transaction
-        let hashed_payload = ProxyContract::hash_payload(&tx.build_for_signing());
+        let hashed_payload = utils::hash_payload(&tx.build_for_signing());
 
         // Log arguments used for signature request
         near_sdk::env::log_str(&format!(
             "Signing request - transaction hash: {:?}, Path: {}, Key Version: {}",
             bs58::encode(&hashed_payload).into_string(),
-            "ed25519:71BRk6DLCgBRrQe6KLyPQGNiXoPJGKhP79Mjyeucu2fN",
+            account_pk_for_mpc,
             0
         ));
         // Create a signature request for the hashed payload
         let request = SignRequest {
             payload: hashed_payload.to_vec(),
-            path: "ed25519:71BRk6DLCgBRrQe6KLyPQGNiXoPJGKhP79Mjyeucu2fN".to_string(), // TODO add this pk value programmatically
+            path: account_pk_for_mpc,
             key_version: 0,
         };
 
@@ -283,48 +258,6 @@ impl ProxyContract {
 
     }
 
-    fn convert_pk_to_omni(pk: &PublicKey) -> omni_transaction::near::types::PublicKey {
-        // TODO We might need to expand this to support ETH/other curve types
-        let public_key_data = &pk.as_bytes()[1..]; // Skipping the first byte which is the curve type
-        //const ED25519_PUBLIC_KEY_LENGTH: usize = 32;
-        const SECP256K1_PUBLIC_KEY_LENGTH: usize = 64;
-        let ed25519_key: [u8; SECP256K1_PUBLIC_KEY_LENGTH] = public_key_data
-            .try_into()
-            .expect("Failed to convert public key");
-
-        OmniPublicKey::SECP256K1(Secp256K1PublicKey::from(ed25519_key))
-    }
-
-    fn hash_payload(payload: &[u8]) -> [u8; 32] {
-        let mut hasher = Sha256::new();
-        hasher.update(payload);
-        hasher.finalize().into()
-    }
-
-    /// Converts a `PublicKey` to a string representation.
-    fn public_key_to_string(public_key: &PublicKey) -> String {
-        let curve_type = public_key.curve_type();
-        let encoded = bs58::encode(&public_key.as_bytes()[1..]).into_string(); // Skipping the first byte which is the curve type
-        match curve_type {
-            CurveType::ED25519 => format!("ed25519:{}", encoded),
-            CurveType::SECP256K1 => format!("secp256k1:{}", encoded),
-        }
-    }
-
-    fn assert_action_allowed(&self, action: &NearAction) {
-        let allowed_contracts = ["wrap.near", "intents.near", "wrap.testnet"];
-        let restricted_methods = ["deposit", "add_public_key"];
-        if !allowed_contracts.contains(&action.contract_id.as_str()) {
-            panic!(
-                "{} is not allowed. Only wrap.near, wrap.testnet, and intents.near are permitted",
-                action.contract_id
-            );
-        }
-        if restricted_methods.contains(&action.method_name.as_str()) {
-            panic!("Method {} is restricted", action.method_name);
-        }
-    }
-
     // Owner methods for managing authorized users
     pub fn add_authorized_user(&mut self, account_id: AccountId) {
         self.assert_owner();
@@ -351,161 +284,6 @@ impl ProxyContract {
             env::predecessor_account_id(),
             self.owner_id,
             "Be gone. You have no power here."
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use near_sdk::{
-        test_utils::{accounts, VMContextBuilder},
-        testing_env,
-    };
-
-    fn get_context(predecessor: AccountId) -> VMContextBuilder {
-        let mut builder = VMContextBuilder::new();
-        builder.predecessor_account_id(predecessor);
-        builder
-    }
-
-    #[test]
-    fn test_new() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-        let contract = ProxyContract::new(accounts(1));
-        assert_eq!(contract.owner_id, accounts(1));
-    }
-
-    #[test]
-    fn test_authorize_user() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-
-        contract.add_authorized_user(accounts(2));
-        assert!(contract.is_authorized(accounts(2)));
-    }
-
-    #[test]
-    fn test_remove_authorized_user() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-
-        contract.add_authorized_user(accounts(2));
-        assert!(contract.is_authorized(accounts(2)));
-
-        contract.remove_authorized_user(accounts(2));
-        assert!(!contract.is_authorized(accounts(2)));
-    }
-
-    #[test]
-    #[should_panic(expected = "Be gone. You have no power here.")]
-    fn test_unauthorized_add_user() {
-        let context = get_context(accounts(2));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-        contract.add_authorized_user(accounts(3));
-    }
-
-    #[test]
-    fn test_get_authorized_users() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-
-        contract.add_authorized_user(accounts(2));
-        contract.add_authorized_user(accounts(3));
-
-        let users = contract.get_authorized_users();
-        assert_eq!(users.len(), 2);
-        assert!(users.contains(&accounts(2)));
-        assert!(users.contains(&accounts(3)));
-    }
-
-    #[test]
-    fn test_set_signer_contract() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-
-        contract.set_signer_contract(accounts(2));
-        assert_eq!(contract.get_signer_contract(), accounts(2));
-    }
-
-    #[test]
-    #[should_panic(expected = "Be gone. You have no power here.")]
-    fn test_unauthorized_set_signer() {
-        let context = get_context(accounts(2));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-        contract.set_signer_contract(accounts(3));
-    }
-
-    #[test]
-    #[should_panic(expected = "Unauthorized: only authorized users can request signatures")]
-    fn test_unauthorized_request_signature() {
-        let context = get_context(accounts(2));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-        contract.request_signature(
-            accounts(3),                       // contract_id: AccountId
-            "test_method".to_string(),         // method_name: String
-            vec![1, 2, 3],                     // args: Vec<u8>
-            near_sdk::json_types::U64(10),                // gas: U64
-            NearToken::from_near(1),           // deposit: NearToken
-            U64(1),                            // nonce: U64
-            Base58CryptoHash::from([0u8; 32]), // block_hash: Base58CryptoHash
-            "secp256k1:abcd1234".to_string(),
-        );
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "danny is not allowed. Only wrap.near, wrap.testnet, and intents.near are permitted"
-    )]
-    fn test_disallowed_action() {
-        //TODO rewrite this as a workspace integration test
-        let context = get_context(accounts(2));
-        testing_env!(context.build());
-        let mut contract = ProxyContract::new(accounts(1));
-
-        testing_env!(get_context(accounts(1)).build());
-        contract.add_authorized_user(accounts(2));
-
-        testing_env!(get_context(accounts(2)).build());
-        contract.request_signature(
-            accounts(3),                       // contract_id
-            "test_method".to_string(),         // method_name
-            vec![1, 2, 3],                     // args
-            near_sdk::json_types::U64(10),                // gas
-            NearToken::from_near(1),           // deposit
-            U64(1),                            // nonce
-            Base58CryptoHash::from([0u8; 32]), // block_hash
-            "secp256k1:abcd1234".to_string(),
-        );
-    }
-
-    #[test]
-    fn test_successful_request_signature() {
-        let context = get_context(accounts(1));
-        testing_env!(context.build());
-
-        let mut contract = ProxyContract::new(accounts(1));
-        contract.add_authorized_user(accounts(2));
-
-        testing_env!(get_context(accounts(2)).build());
-
-        contract.request_signature(
-            "wrap.near".parse().unwrap(),
-            "ft_transfer".to_string(),
-            vec![1, 2, 3],
-            near_sdk::json_types::U64(10),
-            NearToken::from_near(1),
-            U64(1),
-            Base58CryptoHash::from([0u8; 32]),
-            "secp256k1:abcd1234".to_string(),
         );
     }
 }
