+++ 0.0.1/src/scripts/test.sh
@@ -1,2 +1,3 @@
 # Run the tests with output
-cargo test --test proxy_test -- --nocapture
+cargo test --test unit_tests -- --nocapture
+cargo test --test integration_tests -- --nocapture
